+++ src/Components/Common/Header/Header.tsx
@@ -3,72 +3,20 @@ import React from "react";
 import * as S from "./Style";
 import * as I from "../../../Assets";
 import { Button } from "../..";
-import { useState } from "react";
 import SigInModal from "../../SigInModal/SiginModal";
 
 interface HeaderProps {
-  theme: "Login" | "NoneLogin" | "Write";
+  theme: "NoneLogin" | "Login" | "Write";
+  children: React.ReactNode;
 }
 
-const ReturnHeader = (theme: string) => {
-  if (theme === "Login") {
-    return (
-      <>
-        <Button
-          theme="TextButton"
-          fontSize="h5"
-          fontWeight="400"
-          size="Custom"
-          isShadow="No"
-        >
-          새 글 쓰기
-        </Button>
-        <div css={S.ProfileWrapper}>
-          <img src="https://avatars.githubusercontent.com/u/66630940?v=4" />
-          <I.DownArrow />
-        </div>
-      </>
-    );
-  } else if (theme === "NoneLogin") {
-    const [modalState, setModalState] = useState(false);
+const ReturnHeader = (theme: string) => {};
 
-    const closeModal = (e: Event) => {
-      e.preventDefault();
-      setModalState(false);
-    };
-    return (
-      <>
-        <Button
-          theme="OutlinedButton"
-          fontSize="h5"
-          size="Medium"
-          isShadow="No"
-          fontWeight="400"
-          onClick={(closeModal) => setModalState(true)}
-        >
-          로그인
-        </Button>
-
-        <SigInModal modalState={modalState} closeModal={closeModal} />
-      </>
-    );
-  } else if (theme === "Write") {
-    return (
-      <>
-        <div css={S.ProfileWrapper}>
-          <img src="https://avatars.githubusercontent.com/u/66630940?v=4" />
-          <I.DownArrow />
-        </div>
-      </>
-    );
-  }
-};
-
-const Header: React.FC<HeaderProps> = ({ theme }) => {
+const Header: React.FC<HeaderProps> = ({ children, theme }) => {
   return (
     <div css={S.Positioner}>
       <I.PuzzleLogo />
-      <div css={[S.RightContainer[theme]]}>{ReturnHeader(theme)}</div>
+      <div css={[S.RightContainer[theme]]}>{children}</div>
     </div>
   );
 };
