+++ src/Components/Common/Header/Style.ts
@@ -37,10 +37,4 @@ export const RightContainer = {
   `,
 };
 
-export const ProfileWrapper = css`
-  width: 85px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  cursor: pointer;
-`;
+
