+++ src/Components/index.ts
@@ -7,6 +7,7 @@ export { default as Input } from "./Input/Input";
 export { default as ToggleButton } from "./Common/ToggleButton/ToggleButton";
 export { default as ProfileHeader } from "./Profile/ProfileHeader/ProfileHeader";
 export { default as ProfileInfo } from "./Profile/ProfileInfo/ProfileInfo";
+export { default as ProfileCheckHeader } from "./ProfileHeader/ProfileHeader";
 export { default as ProfileSelectBox } from "./Profile/ProfileSelectBox/ProfileSelectBox";
 export { default as WriteSelectItem } from "./Write/WriteSelectItem/WriteSelectItem";
 export { default as WriteTextInput } from "./Write/WriteTextInput/WriteTextInput";
